week-05/DontWreckMyHouse/src/main/java/learn/DontWreckMyHouse/domain/HostService.java
package learn.DontWreckMyHouse.domain;

import learn.DontWreckMyHouse.data.DataException;
import learn.DontWreckMyHouse.data.HostRepository;
import learn.DontWreckMyHouse.models.Host;

import java.util.List;

public class HostService {

    private final HostRepository repository;

    public HostService(HostRepository repository) {
        this.repository = repository;
    }

    public List<Host> findAll() throws DataException {
        return repository.findAll();
    }

    public Host findHostByEmail(String hostEmail) throws DataException {
        return repository.findHostByEmail(hostEmail);
    }
}
